requirements.txt
streamlit
pandas
xgboost
requests
python-dotenv
plotly
scikit-learn
matplotlib
